Phase 1+2: Effects.txt additions, CAttack→CEffect unification, EffectFromName()
Phase 1 (Effects.txt data additions):
- Add Physical Hit effect (HIT, no element, 1d6, range 1)
- Add Poison Bolt effect (HIT, POISON, LINE, 2d8, range 8)
- Add Poison Ball effect (HIT, POISON, BALL, 6d8, radius 3, range 5)
- Add Pick Pocket effect (HIT, TREASURE, range 1) — placeholder for #275

Phase 2 (code: parser prerequisite):
- Extract CEffectDef and CEffect into Effect.h / Effect.cpp
- CAttack reshaped: drop m_dwEffect/m_dwEffectFlags/m_ed; add owned CEffect *m_pEffect
- CAttack::m_dwType (delivery flavor) and m_szDamage preserved unchanged
- Add CDataFile::EffectFromName() — allocates+fills a CEffect from named catalog entry
- Both item Effect and monster Attack named-ref paths now call EffectFromName()
- Old inline Attack format (EFFECT_TYPE_ prefix) also populates m_pEffect
- AIMgr.cpp: m_pCurrentAttack->m_dwEffectFlags → m_pEffect->m_dwFlags with NULL guard
- Monster.h: include Effect.h not Item.h (no CItem/CItemDef usage)
- FileParse.h: include Effect.h not Item.h (CItemDef forward-declared)
- Effect.h: clang-format off guard to preserve JMDefs.h-first include order
