+++ src/AIMgr.cpp
@@ -234,9 +234,8 @@ void CAIBrain::CollideWithPlayer()
         }
 
         float fDamage = m_pParent->Damage( fDamageMult );
-        uint32 dwElement = ( m_pParent->m_pCurrentAttack->m_dwEffectFlags > 0 )
-                               ? (uint32)m_pParent->m_pCurrentAttack->m_dwEffectFlags
-                               : 0;
+        CAttack *pAtk = m_pParent->m_pCurrentAttack;
+        uint32 dwElement = ( pAtk && pAtk->m_pEffect ) ? pAtk->m_pEffect->m_dwFlags : 0;
         g_pGame->GetPlayer()->TakeDamage( fDamage, m_pParent->GetName(), dwElement );
         m_pParent->AttackDone();
     }
